Renamed internet package to better reflect the Tanenbaum model
