+++ src/main/java/layers/network/PacketSender.java
@@ -1,4 +1,4 @@
-package layers.internet;
+package layers.network;
 
 public interface PacketSender {
     public void sendMessage(InternetPacket internetPacket);
